wm: server: track focussed surface
We currently only track the focussed screen, which has mostly been
sufficient until now. Upcoming support for layer-shell and input-inhibit
will require us to keep track of focussed clients which cannot directly
be attributed to a specific screen though.

Introduce a new `tmbr_server` struct member which tracks the currently
focussed surface. Right now, this is always going to be an XDG surface's
surface.

This fixes a bug that when multiple outputs exist where the cursor
resides in a client not part of the current screen, and that client
commits, then the input focus would've been changed to the committing
client.