wm: xdg: damage both old and new area on layout change
When an XDG client gets recalculated, we historically didn't have to
damage its old extents: it would get damaged by the client that now
covers the previously occupied space anyway given that all of the screen
is always occupied.

This has changed with layer-shell support though: if a layer client sets
up an exclusive area, then it could very well be that the resized client
now uncovers an exclusive area. And because we don't damage the old
extents, it would retain its old rendered contents.

Fix the issue by damaging both old and new extents of the XDG client.