wm: xdg: only draw damaged border regions
When drawing borders, we're only drawing those which are contained in
the output damage's region. That's not sufficient though: if the XDG
client is covered by a layer client, then we must make sure to only
redraw border areas which are actually damaged. If we don't, then we may
draw borders over undamaged areas of the layer client.

Fix the issue by only drawing damaged border rectangles.