wm: xdg: skip rendering if not damaged
If the XDG client is not contained by any of the output damage's
regions, we're still going through all of its surfaces and borders. As
all of them do proper damage tracking, this is needless busywork: we
won't draw anything anyway. It can potentially be quite expensive though
because to compute whether anything's damaged, we'll require multiple
allocations.

Fix the issue by skipping all work if the client's extents are not
contained in the output damages.