wm: xdg: reduce reallocations when computing borders
We're currently adding border damages one by one. This is inefficient
given that each addition will require one reallocation.

Fix the issue by instead initializing the border regions with the
client's extents without its extents. We then inverse the area of the
complete client, which leads to the inner region being negated while the
outer borders will be damaged. This lets pixman handle all border
damaging in a single call and thus hopefully optimize the number of
allocations reguired for the damaged regions.