wm: desktop: fix fullscreen reset when changing focus
When changing the desktop focus, we need to reset some properties of the
current client. Most importantly, we need to tell it that it's not
active anymore. But we also need do tell it that it's not the fullscreen
client anymore. But because we first change the focus and only reset
fullscreen mode afterwards, this doesn't work alright currently.

Fix this by first resetting fullscreen mode and then changing the
desktop's focus afterwards.