wm: screen: send frame-done events even if render is skipped
We're currently only sending frame-done events in case we actually draw
something. This has been mostly fine until now and worked alright, but
it's not good practice. Quoting the wayland protocol:

    The server must send the notifications so that a client will not
    send excessive updates, while still allowing the highest possible
    update rate for clients that wait for the reply before drawing
    again. The server should give some time for the client to draw and
    commit after sending the frame callback events to let it hit the
    next output refresh.

As it turns out, not sending the frame event exposes a bug with delayed
frame rendering in case a client is still configuring: if we skip
drawing of a frame because any of the clients hasn't configured yet and
wait until we receive the commit, then it will never arrive because of
the missing frame events.

Fix the issue by sending frame events to both XDG and layer-shell
clients regardless of whether we did render or not.