wm: skip frames when configuring clients
When setting a client's extents, we send it a message to adjust its
size. Resizing takes some time for the client though, and as a result
we tend to render imperfect frames while any client resizes as we're
still rendering its surface with the old extents. These imperfect frames
are quite noticeable, so we should try and avoid rendering them at all.

The best we can do is to track any reconfiguration of client extents,
and if any of the clients hasn't finished configuring then we simply
skip the rendering loop. This is trivial to do by just storing the
serial of the reconfiguration message and then waiting for a commit of
that serial. Until then, rendering is skipped completely such that the
old frame is retained and no imperfect new frame is rendered.

While this strategy is simple enough, we need to account for misbehaving
clients: if a client never finishes to configure, then we'd now be stuck
with the perfect but old frame forever. To avoid this, we set up a
per-client timer which clears the pending configure event after a short
amount of time. This timer is currently set to 50ms -- small enough to
not be a pain in case it fires, but big enough to give the client a few
frames to adjust its extents.

With this in place, we now have a best-effort strategy to avoid
rendering imperfect frames.