wm: xdg: only damage visible clients
Until now, there has been no code paths which could've caused us to
damage an invisible client. This has changed now though with the
introduction of configure timeouts: if a client gets resized and then
gets unmapped before the timeout kicks in for the configuration serial,
then the client wouldn't be part of any desktop in the timeout handler.
But in the timeout handler, we call `tmbr_xdg_client_damage()`, which
unconditionally dereferences the client's desktop pointer. So with above
conditions, this would result in a NULL pointer exception.

Fix the issue by moving the check whether the client is part of a
focussed desktop into the damaging functions.