wm: surface: fix pointer notifications for subsurfaces
When we send focus events for both keyboards and pointers, we need to
treat pointer events specially: instead of always sending them to the
owning parent surface (which is either the XDG or layer-shell surface),
we need to send the event to the subsurface the cursor is currently
pointing at. We don't though, so subsurfaces never get proper pointer
events.

Fix the issue by always sending events to the focussed subsurface, if
one exists.