wm: xdg: implement support for XDG popups
While we're already rendering XDG popups implicitly via the XDG surface
iterators, we do not yet have proper damage tracking for them. As a
result, they only get rendered if their owning XDG surface gets damaged
in the area where the popup is located.

Fix the issue by implementing support for XDG popups. Right now, we only
listen to map and unmap events, which seems to be good enough for now to
properly render them. Missing though is still the need to listen for
commit events, but this can be deferred to a later point.