wm: xdg: only handle focus for visible clients
When a client commits, then we handle focus notifications even if the
client is on no desktop or an inactive one. Only unmapped clients should
ever send commit events, so it shouldn't be possible that we do in fact
end up with a desktop-less client there. And neither should it be
possible for a client to be the focussed surface when its desktop isn't
visible. But making the implicit explicit cannot hurt.

Improve the code by only sending focus notifications in case the client
is on an active desktop.