wm: desktop: explicit reset fullscreen if switching focus
When focussing a different client on a given desktop, then we always
reset the desktop's fullscreen state. We need to be careful though when
focussing the current client to not reset it, which is why we exit early
in `tmbr_desktop_focus_client()` in case we re-focus the current focus.
This hides the fact that we only intend to skip resetting fullscreen
mode in these cases.

Refactor the code to not exit early, but instead conditionally call
`tmbr_desktop_set_fullscreen()` only.