wm: xdg: do not damage whole client after resize
When resizing a client, we wait for it to acknowledge the resize via a
commit event. When it does, we clear the pending configure event and
damage the whole client. Damaging the whole client is not sensible to do
though: we should only apply damages where a surface was actually
damaged in order to accomodate for the resize.

Fix the issue by dropping the damaging part completely. We already do
apply surface damages in the `commit` event handler anyway, so there
should be no need to perform additional damage tracking here.