wm: stop tracking focussed surface
With the introduction of the input-inhibit protocol, we've started to
track two different focusses: first the focussed screen, which
implicitly identifies the focussed client via the focussed desktop. And
second the focussed surface, which either points to the currently
focussed client or to an active inhibitor. This is quite hard to reason
about, and needs us to keep track of who is currently focussed. This is
easy to get wrong, and there's probably cases where it's possible to
override the active inhibitor client.

Improve the code by not tracking the focussed surface at all anymore.
Instead, knowledge about inhibitors is instead put into both
`tmbr_server_find_focus()`, which now returns no focus in case an
inhibitor is active,  and `tmbr_surface_notify_focus()`, which rejects
notifications for any client but the inhibitor if one is active.