wm: screen: fix overlay-layer rendering with fullscreen clients
We currently only render layers when no fullscreen client is active.
While this is perfectly fine for the background, bottom and top layers,
it doesn't make sense for the overlay layer: regardless of whether
there's a fullscreen client or not, it must always get rendered above
everything else.

Fix the issue by always rendering the overlay layer.