meson: Use array to host generated proto files
Generated proto files are only used in a single location: when defining
the resulting executable. We can just as well pass all of the generated
files as a simple array and define this array in `proto/meson.build` so
that the dependent is decoupled from how we assemble the array.

Introduce a new `proto_sources` array hosting all generated proto files.