meson: Reduce boilerplate when generating proto sources
For each of the three non-default Wayland protocols we support, we need
to generate at least two different files with similar commands. This is
repetitive and thus easy to get wrong.

Reduce the boilerplate by converting the code into a `foreach` loop
around all of our protocols with a shared declaration of the custom
targets.