client: Support explicit `--help` parameter
The help is currently always printed when given less than two
parameters. This behaviour is kind of awkward given that it's not
possible to pass `--help` at all. Instead, the user must know to pass
insufficiently many parameters to get the help.

Fix the issue by supporting a `--help` parameter.