meson: Generate version via VCS tag
The version put into the `config.h.in` file is currently always the
project's version. Given that this version does not reflect the case
where there's been additional commits on top of that release, it's
inaccurate whenever timber is built from a non-release commit.

Use the `vcs_tag` custom target to generate version information instead.