wm: xdg: Update pointer focus when moving clients
When moving a client, then it is likely that the current relative cursor
location is invalid after the move. We do not notify the client about
this change though, so it'll behave as if the cursor was still in the
same relative position it was at before the move.

Fix the issue by updating pointer focus when moving an XDG clients'
bounding box.