wm: screen: fix excessive recalculations
In commit c139431 (wm: screen: Adjust to removed scale event,
2021-04-24), we have switched to the newly added `commit` event which
replaced the previous `scale` event in wlroots v0.13. What was missed
though is that the new `commit` event runs not only when the output's
geometry changes, but also when it is for example damaged. As a result,
we're now excessively recalculating the screen's layout whenever any
such unrelated event was fired.

Fix this by conditionally executing the logic with wlroots v0.13 and
newer.