meson: specify minimum required meson version
We do not currently specify a minimum required version for meson. This
both makes failures harder to debug in case one is running a lower
version than that, and furthermore we do not get any warnings when
trying to use features that didn't exist in that version yet.

Set the minimum required meson version to v0.56.2. This is the version
provided by Debian Stable, which is currently the oldest supported
distribution in our CI setup. While at it, make the README more explicit
by detailing the minimum required versions.