ci: use "rolling" release tag for Ubuntu jobs
It's tedious busywork to always keep track of the latest Ubuntu release
and frequently requires work. Ideally, we'd thus switch to the "latest"
build tag instead, which always tracks the latest LTS release. We're not
compatible with that version though: it only provides wlroots v0.10.0,
and we require at least v0.11.0.

Switch our Ubuntu-based jobs to instead use the "rolling" tag, which
points to the latest non-LTS release. Eventually, we should switch this
job to use "latest" when the next LTS version is released. Meanwhile, we
also add a "debian:stable" build tag as the new oldest baseline to
verify we continue working with oldish library versions.