wm: add command to set DPMS state of screens
It is not currently possible to manually set the DPMS state of a
screen. Instead, we always disable all screens after five minutes in
case no input has been registered. This is not configurable though and
thus insufficient.

While we could make this configurable, a simple realization is that a
user can just register for idle timeout events and then run arbitrary
commands on any such idle event. This is as flexible as it can get, and
the functionality is in fact already provided by the swayidle(1)
command. The only thing that's missing though is that users have no way
to modify the DPMS state of a screen.

Implement a new protocol function which allows doing this.