wm: enforce matching client and server version
When client and compositor have mismatching understandings of the custom
control protocol, then havoc may ensue: the client mistakenly invokes
the wrong request or sends invalid request arguments. As a consequence,
we have to ensure that both have the same understanding of the protocol.
This could either be done by adding in backwards compatibility, or more
simply by just enforcing that both use the same protocol version.

Enforce that clients talk the same version as the current protocol
version to fix this source of bugs.