wm: fix announcement of wrong protocol version
The compositor is creating the control global with an incorrect version,
and as a result we never announce the correct version to clients which
want to bind to this global. Furthermore, the protocol compatibility
check in the server was wrong: we now compare that the compositor's
global version is matching the protocol's version, which is wrong.

Fix this by creating the global with the correct version. Drop the
protocol version check: we'll have to reimplement it client-side.