wm: screen: redraw when screen is deselected
When changing to a different screen we only redraw the newly-selected
screen, but never redraw the old now-unselected screen. As a result, the
window border on the old screen is not redrawn and thus looks like it
was still selected.

Fix this by damaging the whole previously-focussed screen on focus
change.