wm: layer: fix adding new clients to wrong screen
When a new layer-shell client is added, then the creator may specify the
output on which it is to be displayed. And while we do set the client's
output as expected, afterwards we fail to respect it and instead add it
to the wrong screen. The result is that it'll not be rendered on the
expected screen, and furthermore the client's state is inconsistent
given that it thinks it belong to a different output than it's being
rendered on.

Fix this bug by adding it to the correct screen.