wm: screen: fix crash due to inconsistent desktop ownership on destroy
Closing layer clients causes us to synchronously execute the unmap
handler, which then causes us to reset the focus to any "normal" XDG
client again. This is also done when destroying an output, but most
notably it happens after we have already moved all desktops to a
different screen. Resetting the focus will thus work on the current
screen, but on a desktop that's not owned by that screen anymore. The
result is that we die when we detect this inconsistency.

Fix this bug by closing layer clients before changing the desktop's
ownership.