wm: screen: fix access of freed memory on destroying screens
When destroying screens we also close layer clients. While this causes
us to execute the `on_unmap` handler synchronously, it will also trigger
the `on_destroy` handler asynchronously. This will thus happen after the
screen has already been fully destroyed and its memory was free'd, but
we're still trying to access the screen in order to recalculate it.

Fix this bug by unsetting the `screen` pointer for layer clients when
the screen is destroyed so that we don't try to recaculate it.