wm: fix hang when reconfiguring disabled display
We're always applying output settings even when we're about to disable
an output. On outputs which have a non-default configuration this means
that we're trying to disable and reconfigure the output at the same
point in time. This is not allowed though and will cause wlroots to
generate an error.

Fix this bug by only applying output configuration when it is supposed
to be enabled.