wm: xdg_client: only rebase cursors when reconfiguring
When a client commits data we're always rebasing the cursor in case it
is currently focussed. This thus happens whenever the client is damaging
its view, which is needlessly excessive: the client hasn't changed its
extents or location, so the relative pointer location cannot have
changed.

Improve this by only rebasing the cursor when we have reconfigured the
view.