wm: server: store renderer in variable
In wlroots v0.15.0 the way the renderer needs to be instantiated has
been changed such that it needs to be created by the server instead of
being created by the backend. We thus need to keep around a variable
which tracks the renderer.

Prepare our code for a fix by storing the renderer in a server's struct
field. Like this, it's easy to retain backwards compatibility.