wm: screen: notify clients of focus loss when switching screens
When switching between screens we first update the focussed screen and
desktop and only call `tmbr_desktop_focus_client()` afterwards. As a
result, the focus as returned by `tmbr_server_find_focus()` will have
changed already, which means that we do not have a chance to detect that
the client focus is in fact changing. And consequentially, the old
client is not getting a notification that it's unfocussed.

Fix this by first focussing the new client and only updating the
server's focus afterwards.