wm: desktop: fix not losing focus when selecting an empty desktop
When an empty desktop is selected then the old client is not properly
losing its focus: we only set the new keyboard and pointer focus in case
there is a client that we're focussing, but we don't ever clear the
focus in case there in no newly focussed client.

Fix this by clearing focus in case we're not focussing a client.