wm: screen: do not damage old screen when changing screen focus
We're damaging the complete old screen in case we're changing the focus
to a different one. This is excessive: we only need to damage the border
of the previously-selected client.

Remove the code which damages the screen. We already damage client
borders when they're unfocussed, so this works as expected already.