wm: xdg-client: fix redrawing border if contents are undamaged
When rendering XDG clients we exit early in case the client isn't
contained in the output's damage. The condition checking for this
doesn't include window borders though, and as a result we fail to redraw
them correctly in case borders are damaged, but contents of the client
are not.

Fix this bug by checking border and contents separately.