wm: desktop: fix focus when changing client focus unsets fullscreen mode
When changing the focus on a desktop which is currently in fullscreen
mode we're resetting the desktop back to tiled mode. As part of this we
need to recalculate the layout, which means that we're also rebasing the
cursor of the currently focussed client.

This rebasing is causing a bug though because of the order in which we
update the focus: we first unfocus any exiting client, focus the new
client, then we unset fullscreen mode and finally we update the desktop
to focus the new client. As a result, we will rebase the cursor for the
still-focussed client when trying to unset its fullscreen mode after we
have notified the new focus of its input focus. The result is that the
internal focus and actual input focus are inconsistent.

Fix this bug by unsetting fullscreen mode before handling the new input
focus.