wm: cursor: fix old client not deactivating when crossing screens
When the cursor crosses screens we need to deactivate the previously
focussed client. This is done by checking whether the currently focussed
client and the newly focussed clients are different ones. And, if so,
then the currently focussed one is deactivated and thus also damaged.

This doesn't work as expected though because we change the focussed
screen before we update the client focus. As a result, we'd thus not
detect the newly focussed client as new anymore because its screen was
selected already.

Fix this by updating the focussed screen after adjusting the client's
focus.