wm: cursor: fix focus when switching to screen without clients
When the cursor is switching to a screen without clients we don't update
the input state correctly because we call no function which would update
the client focus. As a result, the previously focussed node from the old
screen continues to be focussed.

Fix this bug by always setting the desktop focus. In case there is no
client on that desktop we know to unfocus the previously focussed node
as expected.