wm: reintroduce command to handle screen DPMS state
Back when we introduced support for the output-manager protocol we also
removed a bunch of protocols which were used to handle a screen's scale,
resolution and other things which are already provided via the protocol.

As part of this cleanup we also removed DPMS handling because screens
can be enabled and disabled via the protocol. There is a catch though:
the purpose is to _completely_ disable screens, that is they should be
considered to not exist anymore during normal operations. We didn't
treat it like this though, and this is a bug we eventually need to fix.

A consequence of this is that disabling a screen will also cause us to
stop reporting the screen's configuration, and as a result reenabling it
via e.g. wlr-randr will reset the screen to its default configuration.

It is thus clear that it cannot be used as a component to implement
DPMS, as was erroneusly claimed back then. This commit thus reintroduces
the command to handle DPMS for screens.