ci: Drop comment about switching to "ubuntu:latest"
In 9a87f80 (ci: use "rolling" release tag for Ubuntu jobs, 2022-01-08),
we have switched from "ubuntu:21.10" to "ubuntu:rolling". Back then, the
sentiment was that we'd eventually want to switch to the "latest" tag
though, which tracks the latest LTS release in stead of the latest
release overall. Given the frequent backwards-incompatible changes in
wlroots, being compatible with the last two years of its releases would
be quite a bit of work though.

Let's just drop the comment and keep using the "ubuntu:rolling" image.