wm: screen: fix crash on destroy caused by destroying scene nodes
When a screen is torn down, we also destroy its scene nodes. Due to a
misunderstanding, we're destroying both the root scene tree of the
screen as well as all its layer children. This is unnecessary though
because the layer scenes are children of the root scene, which means
they get destroyed when their parent gets destroyed. This then causes
issues when we try to destroy the scene a second time.

Fix this bug by not re-destroying the layer scenes.