wm: client: fix leaking scene tree on destroy
While the XDG client's scene will automatically get destroyed when the
surface is killed, its scene tree is created manually by us and not
tied to the lifetime of the surface. We thus leak its memory when an
XDG client is destroyed.

Fix the leak by destroying the scene tree.