wm: layer: assign layer-client as scene node data
We want to unify the logic to find clients via scenes, which requires us
to be able to derive the client from a scene node. While XDG clients
already set up themselves as `data` pointer of their node, this doesn't
yet happen for layer-shell clients.

Do so as a preparatory step towards code unification.