wm: screen: add screen to output only when it's committed
When adding a new output, we first try to initialize it correctly and
commit its mode. If that fails, we exit early and simply ignore that
screen given that it likely won't be usable for us anyway. But even
though we end up ultimately ignoring that screen, we have in fact
already added it to the output layout.

Fix this by adding the output to the output layout as one of our last
steps.