wm: screen: only set mode if there are any
There are outputs which don't have any modes at all, like CRT monitors.
It's useless to try set and commit any mode for them, and in fact we
already do skip this actino in case we see that there is no preferred
mode. We can skip even more work than this though and also skip the
commit.

Do so by only setting the mode and committing it in case there are any
modes in the first place.