wm: screen: recalculate layout only after it was fully initialized
We're recalculating the screen's layout immediately after we have
initialized its structure. At this point in time we haven't performed
actions like adding the output to the output layout though. While this
is not yet an issue, we'll need to use the output scene's coordinates to
correctly calculate coordinates of scene nodes, which will only be set
to the correct values after the output has been added.

Prepare for this change by recalculating the screen after we have added
its output to the output layout.