wm: screen: fix rendering with multihead setups
When calculating the scene nodes' positions, we always start from the
zero coordinates. The basic assumption is that all scene nodes' coords
are relative to the screen they're drawn on. But this assumption in
wrong: the coordinates are in fact relative to the global output layout.
Consequentially, rendering with multihead setups is severely broken
because every screen would try to render its clients starting at (0,0).

Fix this bug by correctly using coordinates from the output scene.