wm: server: simplify and fix finding clients with multiscreen setups
We're currently manually trying to find a screen at given coordinates
when locating clients, only to then use its scene to locate the client
the cursor is currently pointing at. This is needlessly roundabout: we
can instead just ask the root scene to locate the client for us, which
implicitly searches across the different screens.

Simplify the code to do this. This also fixes an issue with locating
clients caused by the fact that we were using coordinates relative to
the output itself, whereas we are now in fact using coordinates which
are global across al outputs.