wm: state: perform lazy quoting of window titles
The output we return when querying the state of the window manager is
following YAML syntax. In fact, in most cases it works just fine to use
YAML parsers. One important exception though is in case clients have a
title that contains any characters that would be interpreted by YAML
like ampersands.

Fix most of these cases by quoting the title. This is not a proper fix,
but rather a lazy approach because we don't properly escape strings as
we ought to do. It should fix the most common cases though and is good
enough for now.