wm: xdg: fix misbehaving clients due to missing tiling info
Some clients are exhibiting weird behaviour that seems to not play well
with the tiling nature of timber: mpv resizes the window, the GNOME file
chooser displays a bar and has weird geometry, and other such things.

The root cause of this issue seems to be that we never tell clients that
they are indeed tiled, and thus they assume floating window semantics.

Fix this by unconditionally telling all XDG clients that they are tiled.