wm: xdg-client: fix stale screen pointer when reparenting desktops
Starting with 44740e9 (wm: client: move screen into base, 2022-04-29)
the `tmbr_xdg_client` has two references to its screen: once via the
base struct `tmbr_struct`, and once via it's `desktop` pointer. We
don't correctly upate both pointers though when a client's desktop is
being reparented to a different screen. This can lead to a situation
where the owning screen was destroyed, but the client continues to
reference it.

Fix this bug moving the `screen` pointer out of tha base struct again.
Like this, both layer and XDG clients again only have a single pointer
to it, which neatly avoids the problem of meticulusly updating the
screen for XDG clients.