wm: desktop: fix losing input focus when removing clients
When removing clients on a different desktop we will focus neighbouring
clients on the same desktop. When doing that we always adjust the input
focus as well, without taking into account whether the desktop is even
focussed right now. This may cause us to move input focus to a client
that is not visible right now when a client is removed on an inactive
desktop.

Fix this bug by only moving input focus for the currently focussed
desktop.