ci: Drop Coverity static analysis
The Coverity setup is kind of awkward as it is a third-party closed
source static analyis tool. Besides the closedness it also puts some
rather strict requirements on the platform as it really only provides
tools for Debian-based platforms. Furthermore, Coverity has not yet
helped to catch any real bugs in the context of this project.

Let's drop the Coverity job altogther. We have recently added the
infrastructure to run clang-tidy, which provides open-source static
analysis.