wm: screen: stop enabling adaptive sync when adding outputs
Starting with wlroots 0.16.0 and later we fail to correctly initialize
screens. As it turns out this is caused by us enabling adaptive sync
early on.

Skip enabling adaptive sync for the time being. While we could revisit
this at a later point and investigate further, we have added support for
managing adaptive sync via the `wlr_output_management_v1` protocol in
the preceding commits.