wm: redirect subprocess's standard streams to `/dev/null`
When spawning commands we close all standard file descriptors in the
forked process. This is done so that timber and the new process don't
share the same TTY, which makes sense.

There's some commands though that misbehave when they are executed in an
environment that ain't got standard file descriptors set up, with one
recently encountered example being Firefox.

Fix this by redirecting file descriptors to `/dev/null` instead of
closing them.