wm: fix spawning XDG popups with layer-shell as parents
When spawning a new XDG popup we're adding it as a child of either the
XDG surface or layer-shell surface parent. While we do this correctly
for XDG surface parents, spawning popups for layer-shell parents causes
a segfault.

The root cause of this is a simple type confusion: we're passing the
`data` pointer of the `wlr_layer_surface_v1` when creating the XDG scene
tree. But this pointer references the `wlr_scene_layer_surface_v1` and
not the `wlr_scene_tree` as we expect it to.

Fix this bug by correctly passing the scene tree.