wm: fix idle inhibitor not working
In 5851be8 (wm: switch to new idle-notifier-v1 protocol, 2022-11-21),
we've migrated away from the custom KDE idle protocol to the new
idle-notifier-v1 protocol. This commit also simplified the handler for
new input inhibitors to just always set the idle notifier as inhibited.

As part of that we also stopped incrementing the number of inhibitors
that exist, but we of course rely on proper reference counting when idle
inhibitors get destroyed again. Consequentially, tracking of inhibitors
is completely broken.

Let's stop counting inhibitors altogether. Instead, we can simply check
whether the manager still has clients after a destroy was processed.