wm: layer_shell: fix focus notifications for layer clients
When handling motion events for layer-shell clients we first translate
the cursor's coordinates into surface-relative coordinates. While the
logic looks alright on the surface, the logic does not work because we
don't ever assign the layer client's geometry variables. As a result,
things only worked by chance because most layer-shell clients happen to
be rooted at the top-left corner.

Fix this by instead finding the layer-shell client and the subsurface
being pointed at via `tmbr_server_find_client_at()`. While more code,
this also has the benefit that we essentially deduplicate the logic to
determine the position of layer-shell clients and their subsurfaces so
that we only have a single source of truth.