wm: xdg_shell: refactor finding subsurfaces for focus notifications
Refactor finding subsurfaces for focus notifications to work the same as
it does for layer-shell clients by using `tmbr_server_find_client_at()`.
While the resulting code is more verbose, it deduplicates the logic to
find XDG clients and their subsurfaces so that we only have a single
source of truth.