wm: layer_shell: fix handling of negative exclusive zones
Layer-shell clients can specify a negative exclusive zone in order to
tell the compositor that they do not care for any preexisting clients
that have a exclusive zone set. Instead, they should be painted on top
of any such client.

We don't currently treat such clients correctly though when calculating
layers and will process them together with the exclusive clients. In
practice this shouldn't make much of a difference. But let's fix it
anyway.