wm: layer_client: fix focus without subsurface
In 30035f8 (wm: layer_shell: fix focus notifications for layer clients,
2023-04-03), we have refactored the way focus notifications work for
layer-shell clients. This is causing a regression when asked to notify a
layer-shell surface that is not currently pointed to by the cursor or
when not finding its subsurface.

Fix this bug by also focussing the layer-shell client in case we have
not found a layer-shell surface.