wm: only calculate focus notification once on layout change
When recalculating the layout we need to make sure that any client that
is currently focussed will get a notification in case its geometry
changed in any way. This is required so that the client gets an update
on the changed relative cursor position.

This logic is handled when setting a client's geometry: if the geometry
did in fact change, we'll send it a notification in case it's currently
focussed. This is wasteful though because we only call this logic when
the complete desktop layout is recalculated. So what we end up doing is
to repeatedly find the currently focussed node for all clients that are
about to be resized even though only one of them can be focussed anyway.

Fix this inefficiency by instead re-focussing the client once we have
recalculated the complete desktop layout, only.