wm: screen: simplify retrieval of full screen extents
When recalculating the screen layout we need to retrieve the extents and
position of the screen. We do this quite roundabout via the scene output
even though the information would be readily available via the output
layout already.

Refactor the code to instead use the output layout to retrieve the
screen's extents in order to simplify it.