wm/xdg_client: clip surfaces in the scene tree
Some XDG clients don't respect the boundaries we ask them to render to,
which can lead to them to draw into other client's areas. Until now we
couldn't fix this in the scene tree API, but with wlroots 0.17 there now
is a way to clip surfaces.

Use this new functionality so that clients are properly contained now.