wm/xdg_client: mark clients as maximized on older XDG shell versions
The ability to tell XDG clients that they are being tiled has only been
added with the second revision of the XDG shell protocol. Clients that
do not support this version will consequentially not understand those
bits and thus not honor them.

Mark clients that do not know how to handle them to be maximized
instead. In practice, this should have the same effect for clients.