wm/xdg_client: always re-focus clients on commit
When a client is being resized we wait for the resize event to be
acknowledged and then re-focus the client so that the cursor position is
updated accordingly. Furthermore, we skip emitting frames during this
duration so that we don't render partial frames, but timebox it to 50ms
in total.

If the client fails to resize in these 50ms though then we don't update
the pointer focus at all, which causes the pointer location to become
stale. Fix this by instead unconditionally re-focussing clients when
they commit.