global: rename "screen" to "output"
The "screen" terminology has been introduced quite a while ago when
timber was still an X11 window manager. But ever since migrating to
wlroots it has been kind of outdated given that wlroots calls it an
"output" instead. We're thus being inconsistent with most of the
ecosystem out there.

Rename "screen" to "output" to fix this inconsistency.