wm: store wrapped client structs in data pointers
We're currently storing the scene tree nodes in the `data` pointer of
both the XDG and layer-shell client. This has been fine until now
because we never needed anything else. But in the next commit we're
going to need access to other parts of these clients, which makes this
somewhat unhandy.

Refactor the code to instead store pointers to the wrapped client
structures.