wm/output: fix use-after-free on eviction
When tearing down outputs we free all of their desktops and for each of
them, we evict all clients. The iteration over the client's tree causes
a use-after-free though because we'll call `tmbr_tree_find_sibling()` on
clients on tree entries that we have just removed from the tree.

Fix this by manually looping over all clients.