Revert "wm/xdg_client: always re-focus clients on commit"
This reverts commit f168ddadefb5b5d39e15d2accee2accbafaf84c0. This
commit breaks focussing of layer surfaces because we always refocus the
XDG client whenever it commits, which happens basically whenever its
surface gets damaged and thus rather regularly.

The proper fix would be to properly track the focus so that we also know
when we currently focus a layer client. But this is more involed and
requires some additional refactoring to improve focus handling.